Split Function model to Classification and Function models
Refactored data model by dividing Function to Classification and
Function, so that classification tree is separated from handling
processes. So `Function` represents a handling process now, but it
hasn't been renamed in the code or in the API, at least not yet.
Added temporary parent, function_id and name fields to function API
endpoint to maintain backwards compatibility for now.

Refs #48