Cache JHS 191 XML export
Create a new endpoint for “cached” JHS XML export file (the generated xml is saved as a file instead of using Django’s cache.)
The file should be updated every 6 hours in specific times, so create a command for that task.

The endpoint for this cached version of export should be listed in api root and therefore the view is ViewSet type.

Refs #202