Return 401 when requesting function without authentication
Problem is that end-users are linking function by copy&pasting url from
the system to each other. When user is not logged in and opened
function that is not approved the system gave 404. This is causing
confusion between users. We decided to change the error code to 401,
this way UI can ask user to log in or redirect to login page.

Implemented extra checks when retrieving function. These check decide
what code to return. In short return 401 is function is found, but user
can not view it. Due to few corner cases there was need to create few
extra checks. When request includes version and function is not found
return 404.