ElasticSearch Documents
Implement the Document indices for Action, Classification,
Function, Phase and Record models.

The goal is to have the Documents similar so they can be searched
from a single endpoint. Thus they are all based on the
BaseDocument baseclass.

All the attribute keys need to be prefixed with the model name
to distinct the different facets from each other when the
single search enpoint is implemented.

Refs #292