Squash migrations
Some migrations still reference HStoreField but it is no longer
used. Squash the migrations so that new installations can skip
the migrations that use HStoreField so that it is not required
to create the hstore extension for the database.

If the squashed migration is applied, skip tests that are
testing individual migrations since those cannot be backtracked
after squashing.

Refs #301