Convert attribute validation rules to match the general format
A couple of operations are required to convert the JSON generated
by Django admin to be used by the application:

Add a new method to get the attribute validations.
Returns the attribute validation rules with the priority:
1. Type specific attribute validation rules, combined with
the general attribute validation rules.
2. General attribute validation rules if the type specific do
not exist.
3. Hard coded `_attribute_validations` if neither type
specific or general attribute validation rules exist.

Refactor the conditional validation rules from the JSON schema
of the Django admin to the schema supported by the rest of the
application. This needs to be done because the
django-admin-json-editor schema does not support the existing
validation schema.

- Add tests.

Refs #198