Replace PyXB with lxml, refactor JHS exporter
PyXB acts a a blocker for Python and package version updates;
it is no longer maintained, and will stop functioning in
Python >=3.8. Python 3.7 LTS ends this year, and some of
the packages have already dropped 3.7 support.

There were some more recent forks for PyXB, but they
did not seem to have any regular updates, which is why we
ended up with the rewrite solution.