fix: increase uWSGI harakiri timeout
It was reported that some large function update requests in
production are hitting timeouts, so as a quick patch increased uWSGI
harakiri timeout to 175 seconds. To avoid client erroring while
processing is still going on in the backend, the timeout was set to a
lower value than Openshift haproxy timeout 180 seconds (which seems to
be the max value for that).